Added images storage and swipe deleting a single table view row

+++ ToDoListUpdated/ToDoDetailsViewController.swift
@@ -11,18 +11,18 @@ class ToDoDetailsViewController: UIViewController {
     
     var toDoCD : ToDoCD? = nil
     
-
+    
     
     var toDo = ToDo()
-
+    
     @IBOutlet weak var toDoLabel: UILabel!
     
     @IBAction func doneTapped(_ sender: Any) {
-    
+        
         if let context = (UIApplication.shared.delegate as?
-        AppDelegate)?.persistentContainer.viewContext { if let toDo = toDoCD{
-        context.delete(toDo) }
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext() }
+                          AppDelegate)?.persistentContainer.viewContext { if let toDo = toDoCD{
+                              context.delete(toDo) }
+            (UIApplication.shared.delegate as? AppDelegate)?.saveContext() }
         navigationController?.popViewController(animated: true)
     }
     
@@ -31,34 +31,34 @@ class ToDoDetailsViewController: UIViewController {
         super.viewDidLoad()
         
         if let toDo = toDoCD {
-        if toDo.priority == 1{
-            if let name = toDo.name{
-        toDoLabel.text = "🧐 Wichtig ❗️" + name }
+            if toDo.priority == 1{
+                if let name = toDo.name{
+                    toDoLabel.text = "🧐 Wichtig ❗️" + name }
+            }
+            else if toDo.priority == 2{
+                if let name = toDo.name{
+                    toDoLabel.text = " 🤓 Sehr Wichtig ❗️❗️  " + name
+                }
+            }else{
+                if let name = toDo.name{
+                    toDoLabel.text = name }
+                
+                
+                // Do any additional setup after loading the view.
+            }
+            
+            
+            //    @IBAction func doneTapped(_ sender: Any) {
+            //    }
+            /*
+             // MARK: - Navigation
+             
+             // In a storyboard-based application, you will often want to do a little preparation before navigation
+             override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+             // Get the new view controller using segue.destination.
+             // Pass the selected object to the new view controller.
+             }
+             */
+            
         }
-        else if toDo.priority == 2{
-            if let name = toDo.name{
-            toDoLabel.text = " 🤓 Sehr Wichtig ❗️❗️  " + name
-        }
-        }else{
-            if let name = toDo.name{
-        toDoLabel.text = name }
-
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-//    @IBAction func doneTapped(_ sender: Any) {
-//    }
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
     }}
